% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballgown-package.R
\docType{package}
\name{ballgown-package}
\alias{ballgown-package}
\alias{ballgownp}
\title{The ballgown package for analysis of transcript assemblies}
\description{
The ballgown package for analysis of transcript assemblies
}

