% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ballgown}
\alias{ballgown}
\alias{ballgown-package}
\title{constructor function for ballgown objects}
\usage{
ballgown(samples = NULL, dataDir = NULL, samplePattern = NULL,
  bamfiles = NULL, pData = NULL, verbose = TRUE, ...)
}
\arguments{
\item{samples}{vector of file paths to folders containing sample-specific ballgown data
(generated by \code{tablemaker}).  If \code{samples} is provided, \code{dataDir} and
\code{samplePattern} are not used.}

\item{dataDir}{file path to top-level directory containing sample-specific folders with ballgown
data in them.  Only used if \code{samples} is NULL.}

\item{samplePattern}{regular expression identifying the subdirectories of \code{dataDir}
containing data to be loaded into the ballgown object (and only those subdirectories).  Only used
if \code{samples} is NULL.}

\item{bamfiles}{optional vector of file paths to read alignment files for each sample.  If
provided, make sure to sort properly (e.g., in the same order as \code{samples}).  Default NULL.}

\item{pData}{either a \code{data.frame} with rows corresponding to samples and columns
corresponding to phenotypic variables, or a path to a file containing phenotype data.}

\item{verbose}{if \code{TRUE}, print status messages and timing information as the object is
constructed.}

\item{...}{extra arguments for \code{read.table}, if \code{pData} is a path to a file.}
}
\value{
a \code{\link{ballgown}} object
}
\description{
constructor function for ballgown objects

ballgown
}
\details{
Because experimental data is recorded so variably, it is the user's responsibility to
format \code{pData} correctly.  In particular, it's really important that the rows of
\code{pData} (corresponding to samples) are ordered the same way as \code{samples} or the
\code{dataDir}/\code{samplePattern} combo. You can run
\code{file.path(path = dataDir, pattern = samplePattern)} to see the sample order if
\code{samples} was not used.

If you are creating a ballgown object for a large experiment, this function may run slowly and
use a large amount of RAM. We recommend running this constructor as a batch job and saving the
resulting ballgown object as an rda file.  The rda file usually has reasonable size on disk, and
the object in it shouldn't take up too much RAM when loaded, so the time and memory use in
creating the object is a one-time cost.
}
\author{
Leonardo Collado Torres, Alyssa Frazee
}

