% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-data.R
\docType{data}
\name{bg}
\alias{bg}
\title{Toy ballgown object}
\format{a \code{ballgown} object: 100 transcripts, 633 exons, 536 introns}
\description{
Small ballgown object created with simulated toy data, for 
  demonstration purposes
}
\examples{
data(bg)
bg 
# ballgown instance with 100 transcripts and 20 samples
}
\author{
Alyssa Frazee
}
