% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{determine if one set of GRanges fully contains any of another set of GRanges}
\usage{
contains(transcripts, cds)
}
\arguments{
\item{transcripts}{\code{GRangesList} object (assume for now that it
represents transcripts)}

\item{cds}{\code{GRangesList} object (assume for now that it represents sets
of coding sequences)}
}
\value{
vector with length equal to \code{length(transcripts)}, where each
   entry is \code{TRUE} if the corresponding transcript contains a coding
   sequence (i.e., is a superset of at least one entry of \code{cds}).
}
\description{
determine if one set of GRanges fully contains any of another set of GRanges
}
\details{
If \code{gown} is a \code{ballgown} object, \code{transcripts} can
   be \code{structure(gown)$trans} (or any subset).
}
\author{
Alyssa Frazee
}
