% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{data<-}
\alias{data<-}
\alias{data<-,ballgown-method}
\title{Replacement method for data slot in ballgown objects}
\usage{
data(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{data(x)} or a subcomponent}
}
\description{
Replacement method for data slot in ballgown objects
}

