% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{data}
\alias{data}
\alias{data,ballgown-method}
\title{extract data components from ballgown objects}
\usage{
data(x)

\S4method{data}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
list containing elements \code{intron}, \code{exon}, and \code{trans},
 which are feature-by-sample data frames of expression data.
}
\description{
extract data components from ballgown objects
}

