% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dirs}
\alias{dirs}
\alias{dirs,ballgown-method}
\title{extract paths to tablemaker output}
\usage{
dirs(x)

\S4method{dirs}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\description{
extract paths to tablemaker output
}
\examples{
data(bg)
dirs(bg)
}

