% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{expr<-}
\alias{expr<-}
\alias{expr<-,ballgown-method}
\title{Replacement method for expr slot in ballgown objects}
\usage{
expr(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{expr(x)} or a subcomponent}
}
\description{
Replacement method for expr slot in ballgown objects
}

