% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{expr}
\alias{expr}
\alias{expr,ballgown-method}
\title{extract expression components from ballgown objects}
\usage{
expr(x)

\S4method{expr}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
list containing elements \code{intron}, \code{exon}, and 
 \code{trans}, which are feature-by-sample data frames of expression data.
}
\description{
extract expression components from ballgown objects
}
\examples{
data(bg)
names(expr(bg))
class(expr(bg))
dim(expr(bg)$exon)

}
\seealso{
\code{\link{texpr}}, \code{\link{gexpr}}, \code{\link{eexpr}}, 
  \code{\link{iexpr}}
}

