% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gexpr}
\alias{gexpr}
\alias{gexpr,ballgown-method}
\title{extract gene-level expression measurements from ballgown objects}
\usage{
gexpr(x)

\S4method{gexpr}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
gene-by-sample matrix containing per-sample gene FPKMs.
}
\description{
gene-level measurements are done by appropriately combining FPKMs from the transcripts comprising
the gene.
}
\examples{
data(bg)
gene_matrix = gexpr(bg)
}

