% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gffRead}
\alias{gffRead}
\title{read in GTF/GFF file as a data frame}
\usage{
gffRead(gffFile, nrows = -1, verbose = FALSE)
}
\arguments{
\item{gffFile}{name of GTF/GFF on disk}

\item{nrows}{number of rows to read in (default -1, which means read all rows)}

\item{verbose}{if TRUE, print status info at beginning and end of file read. Default FALSE.}
}
\value{
data frame representing the GTF/GFF file
}
\description{
read in GTF/GFF file as a data frame
}
\author{
Kasper Hansen
}
\seealso{
\code{\link{getAttributeField}} to extract data from "attributes" column;
\url{http://useast.ensembl.org/info/website/upload/gff.html} for more information on the GTF/GFF
file format.
}

