% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{indexes<-}
\alias{indexes<-}
\alias{indexes<-,ballgown-method}
\title{Replace method for indexes slot in ballgown objects}
\usage{
indexes(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{indexes(x)} or a subcomponent}
}
\description{
Replace method for indexes slot in ballgown objects
}

