% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{indexes}
\alias{indexes}
\alias{indexes}
\alias{indexes,ballgown-method}
\title{extract the indexes from ballgown objects}
\usage{
indexes(x)

\S4method{indexes}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
list containing elements \code{e2t}, \code{i2t}, \code{t2g}, 
 \code{bamfiles}, and \code{pData}, where \code{e2t} and \code{i2t} are data
 frames linking exons and introns (respectively) to transcripts, \code{t2g}
 is a data frame linking transcripts to genes, and \code{bamfiles} and 
 \code{pData} are described in \code{?ballgown}.
}
\description{
extract the indexes from ballgown objects
}
\examples{
data(bg)
names(indexes(bg))
class(indexes(bg))
head(indexes(bg)$t2g)
}
