% Generated by roxygen2 (4.0.0): do not edit by hand
\name{last}
\alias{last}
\title{get the last element}
\usage{
last(x)
}
\arguments{
\item{x}{anything you can call \code{tail} on (vector, data frame, etc.)}
}
\value{
the last element of \code{x}
}
\description{
get the last element
}
\details{
this function is made of several thousand lines of complex code, so be sure to read it carefully.
}
\author{
Alyssa Frazee
}

