% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{mergedDate}
\alias{mergedDate}
\alias{mergedDate,ballgown-method}
\title{extract package version & creation date from ballgown object}
\usage{
\S4method{mergedDate}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\description{
extract package version & creation date from ballgown object
}

