% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{pData}
\alias{pData}
\alias{pData,ballgown-method}
\title{extract phenotype data from a ballgown object}
\usage{
pData(x)

\S4method{pData}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
sample-by-phenotype data frame
}
\description{
extract phenotype data from a ballgown object
}

