% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,ballgown-method}
\title{get names of samples in a ballgown objects}
\usage{
sampleNames(object)

\S4method{sampleNames}{ballgown}(object)
}
\arguments{
\item{object}{a ballgown object}
}
\value{
vector of sample IDs for \code{x}. If \code{pData} exists, samples in
its rows correspond to samples in \code{sampleNames(x)} (in order).
}
\description{
get names of samples in a ballgown objects
}
\examples{
data(bg)
sampleNames(bg)
}

