% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{simgown}
\alias{simgown}
\title{Toy ballgown object}
\format{a \code{ballgown} object: 100 transcripts, 633 exons, 536 introns}
\description{
Small ballgown object created with simulated toy data, for demonstration purposes
}
\author{
Alyssa Frazee
}

