\name{ss}
\alias{ss}
\title{nice wrapper for string splitting and subsetting}
\usage{
ss(x, pattern, slot = 1, ...)
}
\arguments{
  \item{x}{character vector whose elements are to be split}

  \item{pattern}{what to split by}

  \item{slot}{which slot from each element of x you'd like
  returned (default 1)}

  \item{...}{other arguments to strsplit}
}
\description{
nice wrapper for string splitting and subsetting
}
\details{
internal use only
}
\author{
Andrew Jaffe
}

