% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{structure}
\alias{structure}
\alias{structure}
\alias{structure,ballgown-method}
\title{extract structure components from ballgown objects}
\usage{
structure(x)

\S4method{structure}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
list containing elements \code{intron}, \code{exon}, and 
 \code{trans}. \code{exon} and \code{intron} are \code{GRanges} objects,
 where each range is an exon or intron, and \code{trans} is a
 \code{GRangesList} object, where each \code{GRanges} element is a set of 
 exons representing a transcript.
}
\description{
extract structure components from ballgown objects
}
\examples{
data(bg)
names(structure(bg))
class(structure(bg))
structure(bg)$exon
}
