% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,ballgown-method}
\title{subset ballgown objects to specific samples or genomic locations}
\usage{
subset(x, ...)

\S4method{subset}{ballgown}(x, cond, genomesubset = TRUE)
}
\arguments{
\item{x}{a ballgown object}

\item{cond}{Condition on which to subset. See details.}

\item{genomesubset}{if TRUE, subset \code{x} to a specific part of the genome. Otherwise, subset
x to only include specific samples. TRUE by default.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a subsetted ballgown object, containing only the regions or samples satisfying
  \code{cond}.
}
\description{
subset ballgown objects to specific samples or genomic locations
}
\details{
To use \code{subset}, you must provide the \code{cond} argument as a string representing
a logical expression specifying your desired subset. The subset expression can either involve
column names of \code{texpr(x, "all")} (if \code{genomesubset} is \code{TRUE}) or of
\code{pData(x)} (if \code{genomesubset} is \code{FALSE}). For example, if you wanted a ballgown
object for only chromosome 22, you might call \code{subset(x, "chr == 'chr22'")}.
(Be sure to handle quotes within character strings appropriately).
}
\author{
Alyssa Frazee
}

