% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{texpr}
\alias{texpr}
\alias{texpr,ballgown-method}
\title{extract transcript-level expression measurements from ballgown objects}
\usage{
texpr(x, ...)

\S4method{texpr}{ballgown}(x, meas = "FPKM")
}
\arguments{
\item{x}{a ballgown object}

\item{meas}{type of measurement to extract. Can be "cov", "FPKM", or "all". Default "FPKM".}
}
\value{
transcript-by-sample matrix containing expression values (measured by \code{meas})
}
\description{
extract transcript-level expression measurements from ballgown objects
}

