% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{transcriptIDs}
\alias{transcriptIDs}
\alias{transcriptIDs,ballgown-method}
\title{get numeric transcript IDs from a ballgown object}
\usage{
transcriptIDs(x)

\S4method{transcriptIDs}{ballgown}(x)
}
\arguments{
\item{object}{a ballgown object}
}
\value{
vector of numeric transcript IDs included in the ballgown object
}
\description{
get numeric transcript IDs from a ballgown object
}
\examples{
data(bg)
transcriptIDs(bg)
}

