% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{transcriptNames}
\alias{transcriptNames}
\alias{transcriptNames,ballgown-method}
\title{get transcript names from a ballgown object}
\usage{
transcriptNames(x)

\S4method{transcriptNames}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
vector of transcript names included in the ballgown object. If object
was created using Cufflinks/Tablemaker, these transcript names will be of the
form "TCONS_*". Return vector is named and ordered by corresponding numeric
transcript ID.
}
\description{
get transcript names from a ballgown object
}
\examples{
data(bg)
transcriptNames(bg)
}

