% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{eexpr}
\alias{eexpr}
\alias{eexpr,ballgown-method}
\title{extract exon-level expression measurements from ballgown objects}
\usage{
eexpr(x, meas)

\S4method{eexpr}{ballgown}(x, meas = "rcount")
}
\arguments{
\item{x}{a ballgown object}

\item{meas}{type of measurement to extract. Can be "rcount", "ucount", "mrcount", "cov", "mcov",
or "all". Default "rcount".}
}
\value{
exon-by-sample matrix containing exon-level expression values (measured by \code{meas})
}
\description{
extract exon-level expression measurements from ballgown objects
}

