% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-replace-methods.R
\docType{methods}
\name{expr<-}
\alias{expr<-}
\alias{expr<-}
\alias{expr<-,ballgown-method}
\title{Replacement method for expr slot in ballgown objects}
\usage{
expr(x) <- value

\S4method{expr}{ballgown}(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{expr(x)} or a subcomponent}
}
\description{
Replacement method for expr slot in ballgown objects
}
\examples{
data(bg)
n = ncol(bg@expr$trans)
#multiply all transcript expression measurements by 10:
bg@expr$trans[,11:n] = 10*bg@expr$trans[11:n] 
}
