% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-replace-methods.R
\docType{methods}
\name{indexes<-}
\alias{indexes<-}
\alias{indexes<-}
\alias{indexes<-,ballgown-method}
\title{Replace method for indexes slot in ballgown objects}
\usage{
indexes(x) <- value

\S4method{indexes}{ballgown}(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{indexes(x)} or a subcomponent}
}
\description{
Replace method for indexes slot in ballgown objects
}
\examples{
data(bg)
indexes(bg)$bamfiles = paste0('/path/to/bamfolder/', 
  sampleNames(bg), '_accepted_hits.bam')
}
