% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{mergedDate}
\alias{mergedDate}
\alias{mergedDate,ballgown-method}
\title{extract package version & creation date from ballgown object}
\usage{
mergedDate(x)

\S4method{mergedDate}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\description{
extract package version & creation date from ballgown object
}
\examples{
data(bg)
mergedDate(bg)
}

