% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{pData<-}
\alias{pData<-}
\alias{pData<-,ballgown-method}
\title{Replacement method for pData slot in ballgown objects}
\usage{
pData(x) <- value

\S4method{pData}{ballgown}(x) <- value
}
\arguments{
\item{x}{a ballgown object}

\item{value}{the updated value for \code{pData(x)}.}
}
\description{
Replacement method for pData slot in ballgown objects
}

