% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-replace-methods.R
\docType{methods}
\name{pData<-}
\alias{pData<-}
\alias{pData<-}
\alias{pData<-,ballgown,ANY-method}
\title{Replacement method for pData slot in ballgown objects}
\usage{
pData(object) <- value

\S4method{pData}{ballgown,ANY}(object) <- value
}
\arguments{
\item{object}{a ballgown object}

\item{value}{the updated value for \code{pData(x)}.}
}
\description{
Replacement method for pData slot in ballgown objects
}
\examples{
# add "timepoint" covariate to ballgown object:
data(bg) # already contains pData
pData(bg) = data.frame(pData(bg), timepoint=rep(1:10, 2))
head(pData(bg))
}
