% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{pData}
\alias{pData}
\alias{pData}
\alias{pData,ballgown-method}
\title{extract phenotype data from a ballgown object}
\usage{
pData(object)

\S4method{pData}{ballgown}(object)
}
\arguments{
\item{object}{a ballgown object}
}
\value{
sample-by-phenotype data frame
}
\description{
extract phenotype data from a ballgown object
}
\examples{
data(bg)
pData(bg)
}
