% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ballgown-get-methods.R
\docType{methods}
\name{transcriptIDs}
\alias{transcriptIDs}
\alias{transcriptIDs,ballgown-method}
\title{get numeric transcript IDs from a ballgown object}
\usage{
transcriptIDs(x)

\S4method{transcriptIDs}{ballgown}(x)
}
\arguments{
\item{x}{a ballgown object}
}
\value{
vector of numeric transcript IDs included in the ballgown object
}
\description{
get numeric transcript IDs from a ballgown object
}
\examples{
data(bg)
transcriptIDs(bg)
}

